fix(leif): duplication and extra code
