+++ changelog.d/5-internal/WPB-23789
@@ -6,7 +6,3 @@
 * Library updates:
   - Introduced dedicated error types for ConversationSubsystem to improve error handling
   - Consolidated conversation-related operations that were previously scattered across multiple stores and subsystems
-
-**Note:** The background-worker configuration now has a dependency on the shared global settings.
-Deployments that previously set these values in both galley and background-worker sections
-can consolidate them into the global.settings section.
